refactor: remove internal package; move secrets to own package
