+++ secret/driver/azure.go
@@ -1,4 +1,4 @@
-package secret
+package driver
 
 import (
 	"context"
