+++ secret/driver/base64.go
@@ -1,4 +1,4 @@
-package secret
+package driver
 
 import (
 	"encoding/base64"
